+++ src/asynch.rs
@@ -119,20 +119,11 @@ impl<I: I2c> SHT4x<I> {
         //
         // This is the case even here, where no delay is needed for the
         // sensor to make the data available for reading.
-        #[cfg(feature = "defmt")]
-        defmt::debug!("Reading serial of sensor at {=u8:#02X}", self.address);
 
         self.i2c
             .write(self.address, &[READ_SERIAL_NUMBER_COMMAND])
             .await?;
         self.i2c.read(self.address, &mut self.read_buffer).await?;
-
-        #[cfg(feature = "defmt")]
-        defmt::debug!(
-            "Bytes from sensor {=u8:#02X}: {=[u8; 6]:#02X}",
-            self.address,
-            self.read_buffer
-        );
 
         serial_number_from_read_bytes(Unvalidated::new(self.read_buffer))
     }
@@ -143,9 +134,6 @@ impl<I: I2c> SHT4x<I> {
     ///
     /// An error may be returned if there is a problem with the I2C interface.
     pub async fn soft_reset(&mut self, mut delay: impl DelayNs) -> Result<(), Error<I::Error>> {
-        #[cfg(feature = "defmt")]
-        defmt::debug!("Issuing soft reset to sensor at {=u8:#02X}", self.address);
-
         self.i2c.write(self.address, &[SOFT_RESET_COMMAND]).await?;
         delay.delay_ms(1).await;
         Ok(())
@@ -185,26 +173,9 @@ impl<I: I2c> SHT4x<I> {
         let command = reading_mode.command_byte();
         let us = delay_mode.us_for_reading_mode(reading_mode);
 
-        #[cfg(feature = "defmt")]
-        defmt::debug!(
-            "Measuring from sensor {=u8:#02X}: {} ({=u8:#02X}), {} ({=u32} us)",
-            self.address,
-            reading_mode,
-            command,
-            delay_mode,
-            us
-        );
-
         self.i2c.write(self.address, &[command]).await?;
         delay.delay_us(us).await;
         self.i2c.read(self.address, &mut self.read_buffer).await?;
-
-        #[cfg(feature = "defmt")]
-        defmt::debug!(
-            "Bytes from sensor {=u8:#02X}: {=[u8; 6]:#02X}",
-            self.address,
-            self.read_buffer
-        );
 
         Measurement::from_read_bytes(Unvalidated::new(self.read_buffer))
     }