+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased - 202X-XX-XX
+
+- Removed defmt debug printing.
+
 ## 0.2.0 - 2025-09-18
 
 - Fixed broken rustdoc links in README.